requirements.txt
flask
pywebview
yt-dlp